// src/common/utils/permission.enum.ts
export enum PermissionEnum {
  INVOICES_READ = 'invoices.read',
  INVOICES_WRITE = 'invoices.write',
  PATIENTS_READ = 'patients.read',
  PATIENTS_WRITE = 'patients.write',
  CLINICS_READ = 'clinics.read',
  CLINICS_WRITE = 'clinics.write',
  USERS_READ = 'users.read',
  USERS_WRITE = 'users.write',
  ROLES_READ = 'roles.read',
  ROLES_WRITE = 'roles.write',
  AUDIT_LOG_READ = 'audit.log.read',
  AUDIT_LOG_WRITE = 'audit.log.write'
}